modifed the Grammar and Rule templates to support tracing
it is useful to be able to see a trace of the parser as it
attempts to parse the input using the different rules. this
commit modifies the Grammar and Rule code generator templates
to trace the rules being applied to the input.

an IRuleTrace interface is defined that implies IDisposable
and defines a `Return` method that allows tracers to see the
result of a processed rule. the generated parser rules will
call the `this.RuleTrace` method which must return an object
that implements IRuleTrace. the IRuleTrace object is tracked
with a `using` block so that tracers can reliably manage
entry and exit of each rule.